requirements.txt
asgiref==3.7.2
click==8.1.7
cmake==3.27.7
contourpy==1.2.0
cycler==0.12.1
dj-database-url==2.1.0
Django==4.2.7
django-heroku==0.3.1
django-picklefield==3.1
djangorestframework==3.14.0
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.44.0
gunicorn==21.2.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.1
mpmath==1.3.0
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.3
shapely==2.0.2
six==1.16.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.2.0
typing_extensions==4.8.0
whitenoise==6.6.0